style(cuda): clear the infer-cuda clippy board
- remove dead `..Default::default()` where all fields are set
- elide lifetimes on HcHalf accessors
- match→if let for the single-arm free_retired_fp8_buffers
- iter_mut for the GPTQ zero-fill loop
- drop redundant u64 cast and &reference
- collapse nested if in the TP=1 quant-view path
- remove empty lines after doc comments and outer attributes
- drop the unused DeepSeekV4RopeParameters import

+++ crates/infer-cuda/src/qwen35_lora.rs
@@ -375,7 +375,7 @@ impl Qwen35Model {
             out.push(SharedBf16BaseProjection {
                 layer_idx,
                 proj_suffix: suffix,
-                data_ptr: ptr as u64,
+                data_ptr: ptr,
                 rows: m.rows,
                 cols: m.cols,
             });
